Moving the QTransport abstration layer in his own package.
That allow other packages which need to translate Transport to QTransport to do it.
